requirements.txt
python-telegram-bot[job-queue]
python-dotenv
Pillow
httpx
psycopg2-binary
